start replay, organize tribes controllers
