+++ api/controllers/index.ts
@@ -1,90 +1,93 @@
 import {models} from '../models'
+import * as chats from './chats'
+import * as chatTribes from './chatTribes'
+import * as details from './details'
+import * as contacts from './contacts'
+import * as invites from './invites'
+import * as invoices from './invoices'
+import * as media from './media'
+import * as messages from './messages'
+import * as payments from './payment'
+import * as subcriptions from './subscriptions'
+import * as uploads from './uploads'
+import * as confirmations from './confirmations'
 import {checkTag} from '../utils/gitinfo'
+import * as path from 'path'
 
 const env = process.env.NODE_ENV || 'development';
 console.log("=> env:",env)
-
-let controllers = {
-	messages: require('./messages'),
-	invoices: require('./invoices'),
-	uploads: require('./uploads'),
-	contacts: require('./contacts'),
-	invites: require('./invites'),
-	payments: require('./payment'),
-	details: require('./details'),
-	chats: require('./chats'),
-	subcriptions: require('./subscriptions'),
-	media: require('./media'),
-	confirmations: require('./confirmations')
-}
+const constants = require(path.join(__dirname,'../../config/constants.json'))
 
 async function set(app) {
 
+	const c = await models.Chat.findOne({where:{id:31}})
+	const co = await models.Chat.findOne({where:{id:1}})
+	chatTribes.replayChatHistory(c,co)
+
 	if(models && models.Subscription){
-		controllers.subcriptions.initializeCronJobs()
+		subcriptions.initializeCronJobs()
 	}
 	try{
-		await controllers.media.cycleMediaToken()
+		await media.cycleMediaToken()
 	} catch(e) {
 		console.log('=> could not auth with media server', e.message)
 	}
 
-	app.get('/chats', controllers.chats.getChats)
-	app.post('/group', controllers.chats.createGroupChat)
-	app.post('/chats/:chat_id/:mute_unmute', controllers.chats.mute)
-	app.delete('/chat/:id', controllers.chats.deleteChat)
-	app.put('/chat/:id', controllers.chats.addGroupMembers)
-	app.post('/tribe', controllers.chats.joinTribe)
-	app.put('/group/:id', controllers.chats.editTribe)
+	app.get('/chats', chats.getChats)
+	app.post('/group', chats.createGroupChat)
+	app.post('/chats/:chat_id/:mute_unmute', chats.mute)
+	app.delete('/chat/:id', chats.deleteChat)
+	app.put('/chat/:id', chats.addGroupMembers)
+	app.post('/tribe', chatTribes.joinTribe)
+	app.put('/group/:id', chatTribes.editTribe)
 
-	app.post('/contacts/tokens', controllers.contacts.generateToken)
+	app.post('/upload', uploads.avatarUpload.single('file'), uploads.uploadFile)
 
-	app.post('/upload', controllers.uploads.avatarUpload.single('file'), controllers.uploads.uploadFile)
+	app.post('/invites', invites.createInvite)
+	app.post('/invites/:invite_string/pay', invites.payInvite)
+	app.post('/invites/finish', invites.finishInvite)
 
-	app.post('/invites', controllers.invites.createInvite)
-	app.post('/invites/:invite_string/pay', controllers.invites.payInvite)
-	app.post('/invites/finish', controllers.invites.finishInvite)
+	app.post('/contacts/tokens', contacts.generateToken)
+	app.get('/contacts', contacts.getContacts)
+	app.put('/contacts/:id', contacts.updateContact)
+	app.post('/contacts/:id/keys', contacts.exchangeKeys)
+	app.post('/contacts', contacts.createContact)
+	app.delete('/contacts/:id', contacts.deleteContact)
 
-	app.get('/contacts', controllers.contacts.getContacts)
-	app.put('/contacts/:id', controllers.contacts.updateContact)
-	app.post('/contacts/:id/keys', controllers.contacts.exchangeKeys)
-	app.post('/contacts', controllers.contacts.createContact)
-	app.delete('/contacts/:id', controllers.contacts.deleteContact)
+	app.get('/allmessages', messages.getAllMessages)
+	app.get('/messages', messages.getMessages)
+	app.delete('/message/:id', messages.deleteMessage)
+	app.post('/messages', messages.sendMessage)
+	app.post('/messages/:chat_id/read', messages.readMessages)
+	app.post('/messages/clear', messages.clearMessages)
 
-	app.get('/allmessages', controllers.messages.getAllMessages)
-	app.get('/messages', controllers.messages.getMessages)
-	app.delete('/message/:id', controllers.messages.deleteMessage)
-	app.post('/messages', controllers.messages.sendMessage)
-	app.post('/messages/:chat_id/read', controllers.messages.readMessages)
-	app.post('/messages/clear', controllers.messages.clearMessages)
-
-	app.get('/subscriptions', controllers.subcriptions.getAllSubscriptions)
-	app.get('/subscription/:id', controllers.subcriptions.getSubscription)
-	app.delete('/subscription/:id', controllers.subcriptions.deleteSubscription)
-	app.post('/subscriptions', controllers.subcriptions.createSubscription)
-	app.put('/subscription/:id', controllers.subcriptions.editSubscription)
-	app.get('/subscriptions/contact/:contactId', controllers.subcriptions.getSubscriptionsForContact)
-	app.put('/subscription/:id/pause', controllers.subcriptions.pauseSubscription)
-	app.put('/subscription/:id/restart', controllers.subcriptions.restartSubscription)
+	app.get('/subscriptions', subcriptions.getAllSubscriptions)
+	app.get('/subscription/:id', subcriptions.getSubscription)
+	app.delete('/subscription/:id', subcriptions.deleteSubscription)
+	app.post('/subscriptions', subcriptions.createSubscription)
+	app.put('/subscription/:id', subcriptions.editSubscription)
+	app.get('/subscriptions/contact/:contactId', subcriptions.getSubscriptionsForContact)
+	app.put('/subscription/:id/pause', subcriptions.pauseSubscription)
+	app.put('/subscription/:id/restart', subcriptions.restartSubscription)
 	
-	app.post('/attachment', controllers.media.sendAttachmentMessage)
-	app.post('/purchase', controllers.media.purchase)
-	app.get('/signer/:challenge', controllers.media.signer)
+	app.post('/attachment', media.sendAttachmentMessage)
+	app.post('/purchase', media.purchase)
+	app.get('/signer/:challenge', media.signer)
 
-	app.post('/invoices', controllers.invoices.createInvoice)
-	app.get('/invoices', controllers.invoices.listInvoices)
-	app.put('/invoices', controllers.invoices.payInvoice)
-	app.post('/invoices/cancel', controllers.invoices.cancelInvoice)
+	app.post('/invoices', invoices.createInvoice)
+	app.get('/invoices', invoices.listInvoices)
+	app.put('/invoices', invoices.payInvoice)
+	app.post('/invoices/cancel', invoices.cancelInvoice)
 
-	app.post('/payment', controllers.payments.sendPayment)
-	app.get('/payments', controllers.payments.listPayments)
+	app.post('/payment', payments.sendPayment)
+	app.get('/payments', payments.listPayments)
 
-	app.get('/channels', controllers.details.getChannels)
-	app.get('/balance', controllers.details.getBalance)
-	app.get('/balance/all', controllers.details.getLocalRemoteBalance)
-	app.get('/getinfo', controllers.details.getInfo)
-	app.get('/logs', controllers.details.getLogsSince)
-	app.get('/info', controllers.details.getNodeInfo)
+	app.get('/channels', details.getChannels)
+	app.get('/balance', details.getBalance)
+	app.get('/balance/all', details.getLocalRemoteBalance)
+	app.get('/getinfo', details.getInfo)
+	app.get('/logs', details.getLogsSince)
+	app.get('/info', details.getNodeInfo)
 
 	app.get('/version', async function(req,res) {
 		const version = await checkTag()
@@ -112,4 +115,22 @@ const login = (req, res) => {
 	}
 }
 
-export {set, controllers}
+const msgtypes = constants.message_types
+const ACTIONS = {
+    [msgtypes.contact_key]: contacts.receiveContactKey,
+    [msgtypes.contact_key_confirmation]: contacts.receiveConfirmContactKey,
+    [msgtypes.message]: messages.receiveMessage,
+    [msgtypes.invoice]: invoices.receiveInvoice,
+    [msgtypes.direct_payment]: payments.receivePayment,
+    [msgtypes.confirmation]: confirmations.receiveConfirmation,
+    [msgtypes.attachment]: media.receiveAttachment,
+    [msgtypes.purchase]: media.receivePurchase,
+    [msgtypes.purchase_accept]: media.receivePurchaseAccept,
+    [msgtypes.purchase_deny]: media.receivePurchaseDeny,
+    [msgtypes.group_create]: chats.receiveGroupCreateOrInvite,
+    [msgtypes.group_invite]: chats.receiveGroupCreateOrInvite,
+    [msgtypes.group_join]: chats.receiveGroupJoin,
+    [msgtypes.group_leave]: chats.receiveGroupLeave,
+}
+
+export {set, ACTIONS}
